Add `clone()` function to copy color object
This function uses `this.rgb()` to "serialize" the color object and pass
it to the Color constructor to create a new color object with values
equal to the original.

The new function is documented in the readme.

There are tests to make sure the values of the cloned object match the
original and that if the cloned object is manipulated the original
object does not change.